feat(share): single-asset links open the same review screen as folder links
Closes #117 and #123. Both had the same root cause and #123 named the fix:
converge the single-asset path onto the folder path's stack rather than
extending the bespoke player.

A share link targeting one asset rendered ShareViewer -> ShareMediaViewer
plus GuestCommentInput: a plain <video controls>, a textarea that posted
only {body, guest_email, guest_name}, and no version concept at all. The
same asset shared inside a folder rendered ReviewProvider, VideoPlayer,
CommentInput and VersionSwitcher, so it had timecode-attached comments,
annotations, and version-scoped streams and comments. Identical content
behaved differently depending on how it was shared.

ShareReviewScreen, ShareReviewInner and GuestIdentityPrompt moved out of
folder-share-viewer into share-review-screen.tsx and are now used by both
paths. The move is clean: those three referenced no helper local to that
module. handleDownload and triggerDownload went to share-download.ts,
because importing them back from folder-share-viewer would have made the
two modules circular.

onBack is optional now and the back button is gated on it, since a
single-asset link has no grid to return to.

The bespoke tree is deleted rather than left sitting there: ShareViewer,
ShareMediaViewer, ShareTopBar, GuestCommentList, GuestCommentItem,
GuestApprovalActions, guest-comment-input.tsx, and the types only they
used. That is what the net -937 lines is.

Two behaviour changes I did not want buried in a refactor.

Guest identity: the single-asset path stored it under
freeframe_guest_identity and the shared screen uses ff_guest_identity, so
a guest who had identified themselves on such a link is asked once more.

Branding: ShareTopBar showed the instance logo and share name above the
player and nothing in the shared screen does, so a single-asset link no
longer displays it. The folder path's asset view never did either, so
this is now consistent rather than a regression against parity, but it is
a loss for white-labelled instances and is worth fixing for both paths
rather than reinstating for one. Filed separately.

That also orphaned prefersLightMarkOn, added this morning for the mark on
ShareTopBar's accent circle, so it and its tests go with the component.

Verified against a real share link on the dev stack: version switcher,
timecode chip, annotation controls and composer all present, back button
correctly absent, guest versions and comments endpoints both 200, console
clean.
